+++ example/android/app/src/main/kotlin/com/example/popup_menu/MainActivity.kt
@@ -1,0 +1,6 @@
+package com.example.popup_menu
+
+import io.flutter.embedding.android.FlutterActivity
+
+class MainActivity: FlutterActivity() {
+}