+++ test/popup_menu_test.dart
@@ -1,9 +1,7 @@
 import 'package:flutter_test/flutter_test.dart';
 
-import 'package:popup_menu/popup_menu.dart';
-
 void main() {
-  test('adds one to input values', () {
-    
+  testWidgets('popup menu ...', (tester) async {
+    // TODO: Add test for popup menu
   });
 }